// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  EDITOR: 'EDITOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ContentStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ContentStatus = (typeof ContentStatus)[keyof typeof ContentStatus]


export const BookingStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentGateway = {
  ESEWA: 'ESEWA',
  KHALTI: 'KHALTI',
  STRIPE: 'STRIPE',
  BANK_TRANSFER: 'BANK_TRANSFER'
} as const

export type PaymentGateway = (typeof PaymentGateway)[keyof typeof PaymentGateway]


export const EnquiryType = {
  CONTACT: 'CONTACT',
  PROPOSE: 'PROPOSE'
} as const

export type EnquiryType = (typeof EnquiryType)[keyof typeof EnquiryType]


export const SectionType = {
  HERO: 'HERO',
  SEARCH: 'SEARCH',
  FEATURED_PACKAGES: 'FEATURED_PACKAGES',
  DESTINATIONS: 'DESTINATIONS',
  TESTIMONIALS: 'TESTIMONIALS',
  BLOG_PREVIEW: 'BLOG_PREVIEW',
  RICH_TEXT: 'RICH_TEXT',
  GALLERY: 'GALLERY',
  CTA: 'CTA',
  STATS: 'STATS',
  CONTACT_FORM: 'CONTACT_FORM',
  PAGE_HEADER: 'PAGE_HEADER',
  FEATURE_GRID: 'FEATURE_GRID',
  CHECKLIST: 'CHECKLIST',
  CREDENTIALS: 'CREDENTIALS',
  IMAGE_BAND: 'IMAGE_BAND'
} as const

export type SectionType = (typeof SectionType)[keyof typeof SectionType]